backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai
python-dotenv
python-multipart